Rename io.aims.output to io.aims.outputs
Keep it consistent with inputs naming scheme
